ai-recipe-shoplist/requirements_freezen.txt
aiofiles==23.2.1
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
astroid==4.0.1
attrs==25.4.0
autoflake==2.3.1
azure-core==1.36.0
azure-identity==1.25.1
beautifulsoup4==4.14.2
black==23.12.1
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dill==0.4.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.119.0
flake8==6.1.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
isort==5.13.2
Jinja2==3.1.6
jiter==0.11.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
msal==1.34.0
msal-extensions==1.3.1
mypy==1.18.2
mypy_extensions==1.1.0
numpy==1.26.4
ollama==0.6.0
openai==1.109.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.5.0
playwright==1.55.0
pluggy==1.6.0
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyee==13.0.0
pyflakes==3.1.0
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
pylint==4.0.1
pymemcache==4.0.0
PyNaCl==1.6.0
PySocks==1.7.1
pytest==7.4.4
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
selenium==4.37.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
starlette==0.48.0
tenacity==8.5.0
tiktoken==0.12.0
tomlkit==0.13.3
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.2.0
